+++ terratex_reallife/SQLS/update.1.2.0.sql
@@ -1,0 +1,1 @@
+ALTER TABLE mta_terratex_reallife.user_gold ADD SkillBooster INT DEFAULT 0 NULL;